Fix email confirmation. Add competition API tests.
