+++ client/config/webpack/production.js
@@ -75,7 +75,6 @@ module.exports = {
       'process.env.NODE_ENV': '"production"',
       NODE_ENV: '"production"',
       API_BASE_URL: `"${process.env.API_BASE_URL}"`,
-      COMPETITION_BASE_URL: `"${process.env.COMPETITION_BASE_URL}"`,
       APP_ID: `"${process.env.APP_ID}"`,
     }),
     new webpack.optimize.UglifyJsPlugin({ minimize: true }),
